requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aioice==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
av==13.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
fastapi==0.115.12
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-genai==1.20.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h264-profile-level-id==1.0.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ifaddr==0.2.0
jiter==0.10.0
joblib==1.5.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
mcp==1.9.4
mdurl==0.1.2
msgpack==1.1.1
multidict==6.4.4
numba==0.61.2
numpy==2.2.6
openai==1.88.0
packaging==25.0
pillow==10.4.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyee==11.1.0
pygments==2.19.1
pyjwt==2.10.1
pylibsrtp==0.12.0
pyopenssl==25.1.0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.31.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
sdp-transform==1.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.3.6
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
twilio==9.6.3
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
videosdk==0.1.0
videosdk-agents==0.0.14
videosdk-plugins-google==0.0.6
videosdk-plugins-openai==0.0.8
vonage==4.4.3
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.7
vonage-numbers==1.0.4
vonage-sms==1.1.6
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.4.0
vsaiortc==0.0.8
websockets==15.0.1
yarl==1.20.1
